requirements.txt
pandas
numpy
requests
psycopg2-binary
pytz
